inference-cloud-function/requirements.txt
torch
torchvision
pytorch_lightning
zarr
pandas
tqdm
gcsfs
google-cloud-bigquery
google-cloud-storage
db-dtypes
google-cloud-pubsub